added unit tests; added http interceptor to convert keys to camelCase
The aptly API is returning responses that have PascalCase keys. This
is causing major issues with the model interfaces.

Client code must access API data using PascalCase but the models are
defined with camelCase member names. This in turn causes the TypeScript
compiler to complain about non-existent properties.

To cope, these models were being passed around as any -- defeating
their purpose. This interceptor should allow client to code to use
the model definitions.

It would be better still if the swagger code-gen could be configured
to honor the capitalization of resource definitions.